renamed files to have a more descriptive name, and updated the code to reflect the new names.
