+++ Opt_SSPG.py
